+++ packages/utils/src/client/fetchEventSource/index.ts
