+++ packages/utils/src/client/fetchEventSource/parse.ts
