+++ packages/utils/src/client/index.ts
@@ -1,4 +1,5 @@
 export * from './clipboard';
 export * from './downloadFile';
 export * from './exportFile';
+export * from './fetchEventSource';
 export * from './sanitize';
