+++ src/services/chat/contextEngineering.test.ts
@@ -260,17 +260,13 @@ describe('contextEngineering', () => {
       },
     ];
 
-    const result = await contextEngineering(
-      {
-        messages,
-        model: 'gpt-4',
-        provider: 'openai',
-      },
-      {
-        isWelcomeQuestion: true,
-        trace: { sessionId: 'inbox' },
-      },
-    );
+    const result = await contextEngineering({
+      messages,
+      model: 'gpt-4',
+      provider: 'openai',
+      isWelcomeQuestion: true,
+      sessionId: 'inbox',
+    });
 
     // Should have system message with inbox guide content
     const systemMessage = result.find((msg) => msg.role === 'system');
@@ -295,16 +291,12 @@ describe('contextEngineering', () => {
       },
     ];
 
-    const result = await contextEngineering(
-      {
-        messages,
-        model: 'gpt-4',
-        provider: 'openai',
-      },
-      {
-        historySummary,
-      },
-    );
+    const result = await contextEngineering({
+      messages,
+      model: 'gpt-4',
+      historySummary,
+      provider: 'openai',
+    });
 
     // Should have system message with history summary
     const systemMessage = result.find((msg) => msg.role === 'system');
